+++ src/validateInvocation.ts
@@ -30,6 +30,6 @@ export default async function validateInvocation(
     });
   }
 
-  const apiClient = createAPIClient(config);
+  const apiClient = createAPIClient(config, context.logger);
   await apiClient.verifyAuthentication();
 }